bms/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pytest==7.4.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pytest-cov==4.1.0
flake8==6.1.0
pylint==3.0.2
aiohttp==3.9.0
asyncio==3.4.3
selenium==4.15.2
structlog==23.2.0